currentLog: ILog is not in global state. need to implement editing of a single log as well as get rid of createdAt property

+++ frontend/src/components/Actions/Log.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import axios, { AxiosResponse } from 'axios';
-import { Delete as DeleteIcon } from '@material-ui/icons';
+import { Delete as DeleteIcon, Edit as EditIcon } from '@material-ui/icons';
 import { IconButton } from '@material-ui/core';
-import { getToken } from '../../context/GlobalState';
+import { getToken, GlobalContext } from '../../context/GlobalState';
 import {
   HandleLogActionType,
   ILog,
@@ -10,26 +10,40 @@ import {
   MessageType,
   SetAlertsAndHandleResponseType,
 } from '../../type';
+import { Logs } from '../../enums';
 
 const Log: React.FC<{
   setAlertsAndHandleResponse: SetAlertsAndHandleResponseType;
   row: ITask | ILog;
 }> = (props) => {
-  const handleAction: HandleLogActionType = async (e, logItemId) => {
+  const { dispatch } = useContext(GlobalContext);
+
+  const handleAction: HandleLogActionType = async (e, logItemId, command) => {
     try {
-      const res: AxiosResponse<MessageType> = await axios.delete(
-        `api/task/log/${logItemId}`,
-        {
-          headers: { Authorization: `Bearer ${getToken()}` },
-        }
-      );
-      props.setAlertsAndHandleResponse(
-        'delete',
-        res.data.message,
-        'logs',
-        props.row.projectNumber,
-        null
-      );
+      if (command === 'delete') {
+        const res: AxiosResponse<MessageType> = await axios.delete(
+          `api/task/log/${logItemId}`,
+          {
+            headers: { Authorization: `Bearer ${getToken()}` },
+          }
+        );
+        props.setAlertsAndHandleResponse(
+          'delete',
+          res.data.message,
+          'logs',
+          props.row.projectNumber,
+          null
+        );
+      } else {
+        const res: AxiosResponse<ILog> = await axios.get(
+          `/api/task/singleLog/${props.row.id}`,
+          {
+            headers: { Authorization: `Bearer ${getToken()}` },
+          }
+        );
+
+        dispatch({ type: Logs.setLog, payload: res.data });
+      }
     } catch (err) {
       props.setAlertsAndHandleResponse(
         'error',
@@ -42,7 +56,12 @@ const Log: React.FC<{
   };
   return (
     <>
-      <IconButton onClick={(e) => handleAction(e, props.row.id)}>
+      <IconButton
+        onClick={(e) => handleAction(e, props.row.projectNumber, 'edit')}
+      >
+        <EditIcon />
+      </IconButton>
+      <IconButton onClick={(e) => handleAction(e, props.row.id, 'delete')}>
         <DeleteIcon />
       </IconButton>
     </>
